backend/requirements.txt
# Core FastAPI
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2

# Security
passlib==1.7.4
argon2-cffi==23.1.0
authlib==1.6.0
python-dotenv==1.1.0
email-validator==2.2.0

# HTTP Clients
httpx==0.28.1

# Database
asyncpg==0.30.0
psycopg2-binary==2.9.9

# AI/ML
google-generativeai==0.4.0

# Testing
pytest==8.4.0
pytest-asyncio==1.0.0
